Refactoring: extract dump loading
